Added sections editing, more gui elements

+++ src/log/charter/song/Position.java
@@ -3,7 +3,7 @@ package log.charter.song;
 public class Position implements Comparable<Position> {
 	public double pos;
 
-	public Position(final long pos) {
+	public Position(final double pos) {
 		this.pos = pos;
 	}
